% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_count_pct}
\alias{func_count_pct}
\title{Format Count and Percentage}
\usage{
func_count_pct(count, pct, dec)
}
\arguments{
\item{count}{The count.}

\item{pct}{The percentage.}

\item{dec}{The number of decimal places for the percentage.}
}
\value{
A character string with the formatted count and percentage.
}
\description{
Format Count and Percentage
}
