% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_dx_clean_na}
\alias{func_dx_clean_na}
\title{Basic Diagnosis Code Field Cleaner}
\usage{
func_dx_clean_na(x)
}
\arguments{
\item{x}{The character string (e.g., variable) containing the diagnosis code values.}
}
\value{
A character string with the cleaned values of the diagnosis code.
}
\description{
This function capitalizes the a diagnosis code variable, removes any decimals,
removes leading and trailing blanks, and will convert any "" to NA.
}
\examples{
# Basic usage
func_dx_clean_na(" q25.1  ")
func_dx_clean_na("q25.1a  ")
func_dx_clean_na("")

# Using across multiple diagnosis code variables in a dataset
# Original data
df <- tibble::tibble(dx1=" q25.1  ", dx2="q25.1a  ", dx3="", dx4=NA)
df
# Applying the cleaning function
df <- df |>
  dplyr::mutate(
    across(.cols = num_range("dx", 1:4),
           .fns  = ~ func_dx_clean_na(.x) )
  )
# Printing the cleaned dataset
df
}
