% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_dx_scan_count}
\alias{func_dx_scan_count}
\title{Create Counts of Diagnostic Conditions Based on Diagnosis Code Variables}
\usage{
func_dx_scan_count(df, codes)
}
\arguments{
\item{df}{The data frame which contains the diagnosis code variables.}

\item{codes}{A regular expression (pattern) reflecting the codes of interest.}
}
\value{
A numeric value of 0 or 1.
}
\description{
This function will scan diagnosis code variables in a data frame for specific codes
the user specifies and will create a variable reflecting the number of variables
which contained desired codes reflected in a code pattern (a count).
}
\examples{
# Original data
df <- tibble::tibble(
  id  = c(1, 2, 3),
  dx1 = c("X34.5", "X35.6", "X36.7"),
  dx2 = c("G33.1", "G34.3", "G35.4"),
  dx3 = c("F25.4", "F26.7", "F27.8"),
  dx4 = c(NA,      "H44.8", "H45.9"),
  dx5 = c(NA,      NA,      "I56.0")
)
df
# Create three new variables for each observation
# cnt_any_g_code should count the number of diagnoses that begin with "G"
# cnt_g33_code should count the number of diagnoses that begin with "G33"
# cnt_code_ends_even should count the number of diagnoses that end in an even number
df <- df |>
  dplyr::mutate(
    cnt_any_g_code     = func_dx_scan_count(df |> dplyr::select(dx1:dx5), "^G"),
    cnt_g33_code       = func_dx_scan_count(df |> dplyr::select(dx1:dx5), "^G33"),
    cnt_code_ends_even = func_dx_scan_count(df |> dplyr::select(dx1:dx5), "[02468]$")
  )
df
}
