% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_myround}
\alias{func_myround}
\title{Round Numbers with Commas}
\usage{
func_myround(num, dec)
}
\arguments{
\item{num}{The number to be rounded.}

\item{dec}{The number of decimal places.}
}
\value{
A character string of the formatted number.
}
\description{
Round Numbers with Commas
}
