% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_addci}
\alias{func_addci}
\title{Add Confidence Intervals}
\usage{
func_addci(est, low, hi, dec)
}
\arguments{
\item{est}{The estimate.}

\item{low}{The lower bound of the interval.}

\item{hi}{The upper bound of the interval.}

\item{dec}{The number of decimal places.}
}
\value{
A character string with the estimate and confidence intervals.
}
\description{
Add Confidence Intervals
}
