% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_myround}
\alias{func_myround}
\title{Round Numbers with Commas}
\usage{
func_myround(num, dec)
}
\arguments{
\item{num}{The number to be rounded.}

\item{dec}{The number of decimal places.}
}
\value{
A character string of the formatted number.
}
\description{
This function rounds a number to a specified number of decimal places and
formats it with commas as thousand separators.
}
\examples{
# Basic usage
func_myround(12345.6789, 2)

# Using with dplyr to create a new formatted variable
library(dplyr)
df <- tibble::tibble(value = c(12345.6789, 98765.4321))
df <- df |> dplyr::mutate(formatted_value = func_myround(value, 2))
}
